fix: add fizz.env

+++ release/build_release.sh
@@ -40,6 +40,7 @@ for PLATFORM in "${PLATFORMS[@]}"; do
     # Include the shell script only for macOS and Linux
     if [[ "$PLATFORM" != windows* ]]; then
         cp "$PROJECT_DIR/fizz" "$TARGET_DIR"
+        cp "$SCRIPT_DIR/fizz.env" "$TARGET_DIR"
     fi
 
     # Create archives
@@ -49,8 +50,11 @@ for PLATFORM in "${PLATFORMS[@]}"; do
         tar -czf "releases/$TARGET_DIR.tar.gz" "$TARGET_DIR"
     fi
 
-    echo "Packaged: $TARGET_DIR"
-    rm -rf "$TARGET_DIR"  # Cleanup
+    # if cleanup the target dir if DISABLE_CLEANUP is not set
+    if [[ "${DISABLE_CLEANUP}" != true ]]; then
+      echo "Packaged: $TARGET_DIR"
+      rm -rf "$TARGET_DIR"  # Cleanup
+    fi
 done
 
 echo "All builds completed. Archives are in the releases/ directory."
